pyproject.toml
[project]
name = "brain-occlusion-detection"
version = "0.1.0"
description = "Brain occlusion detection using deep learning with medical imaging (NIfTI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "torchaudio>=2.5.0",
    "monai>=1.3.0",
    "nibabel>=5.0.0",
    "simpleitk>=2.3.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tensorboard>=2.15.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.3.0",
    "vit-pytorch>=1.17.1",
    "pillow>=12.0.0",
    "pandas>=2.3.3",
    "timm>=1.0.22",
    "seaborn>=0.13.2",
    "antspyx>=0.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu124", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
